Move early-stage rewrites to a new module, jsparagus.rewrites.
