+++ tests/test.py
@@ -6,7 +6,7 @@ import unittest
 import typing
 
 import jsparagus
-from jsparagus import gen, lexer
+from jsparagus import gen, lexer, rewrites
 from jsparagus.grammar import (Grammar, Production, CallMethod, Nt,
                                Optional, LookaheadRule, NtDef, Var)
 from js_parser.parse_esgrammar import parse_esgrammar
@@ -390,18 +390,18 @@ class GenTestCase(unittest.TestCase):
     def testExpandOptional(self):
         grammar = Grammar({'goal': [[]]})
         empties = {}
-        # Unit test for gen.expand_optional_symbols_in_rhs
+        # Unit test for rewrites.expand_optional_symbols_in_rhs
         self.assertEqual(
-            list(gen.expand_optional_symbols_in_rhs(['ONE', 'TWO', '3'],
-                                                    grammar, empties)),
+            list(rewrites.expand_optional_symbols_in_rhs(['ONE', 'TWO', '3'],
+                                                         grammar, empties)),
             [(['ONE', 'TWO', '3'], {})])
         self.assertEqual(
-            list(gen.expand_optional_symbols_in_rhs(
+            list(rewrites.expand_optional_symbols_in_rhs(
                 ['a', 'b', Optional('c')], grammar, empties)),
             [(['a', 'b'], {2: None}),
              (['a', 'b', 'c'], {})])
         self.assertEqual(
-            list(gen.expand_optional_symbols_in_rhs(
+            list(rewrites.expand_optional_symbols_in_rhs(
                 [Optional('a'), Optional('b')], grammar, empties)),
             [([], {0: None, 1: None}),
              (['a'], {1: None}),
